relocated fpo test class
